Transaction View, SO Entry and PO Entry fixes
PO Entry
- the Add Row and Clear Rows buttons now function
- Entering a PO now adds the correct amounts to the On Order column for
the specified Stock items
SO Entry
- Entering a Sale now removes the correct amount in the Amount column
for the specified Stock items.
Transaction View
- Deleting transaction (Purchase or Sale) will now adjust the Stock item
Amount and On Order columns appropriately from the values recorded in
the Purchase Order Line and Sale Order Line Tables

Current Bugs:
Transaction View
- The total shown is inaccurate.  It doesn't factor in the amounts of
each item bought, so the total displays the same for buying 1 item as it
would for buying 5.